+++ packages/website/src/App.tsx
@@ -14,6 +14,7 @@ import Faq from "./pages/docs/Faq";
 import ManualTesting from "./pages/docs/ManualTesting";
 import Solana from "./pages/docs/Solana";
 import Evm from "./pages/docs/Evm";
+import Iac from "./pages/docs/Iac";
 
 export function App() {
   return (
@@ -32,6 +33,7 @@ export function App() {
           <Route path="manual-testing" element={<ManualTesting />} />
           <Route path="solana" element={<Solana />} />
           <Route path="evm" element={<Evm />} />
+          <Route path="iac" element={<Iac />} />
         </Route>
         <Route path="*" element={<NotFound />} />
       </Route>